+++ test/unit/control/controller.py
@@ -12,7 +12,7 @@ import stem.socket
 import stem.util.system
 import stem.version
 
-from stem import InvalidArguments, InvalidRequest, ProtocolError, UnsatisfiableRequest
+from stem import ControllerError, InvalidArguments, InvalidRequest, ProtocolError, UnsatisfiableRequest
 from stem.control import _parse_circ_path, Listener, Controller, EventType
 from stem.exit_policy import ExitPolicy
 from test import mocking
@@ -22,6 +22,8 @@ try:
   from unittest.mock import Mock, patch
 except ImportError:
   from mock import Mock, patch
+
+NS_DESC = 'r %s %s u5lTXJKGsLKufRLnSyVqT7TdGYw 2012-12-30 22:02:49 77.223.43.54 9001 0\ns Fast Named Running Stable Valid\nw Bandwidth=75'
 
 
 class TestControl(unittest.TestCase):
@@ -380,6 +382,35 @@ class TestControl(unittest.TestCase):
     self.assertEqual(432, self.controller.get_pid())
 
   @patch('stem.control.Controller.get_info')
+  def test_get_network_status_for_ourselves(self, get_info_mock):
+    """
+    Exercises the get_network_status() method for getting our own relay.
+    """
+
+    # when there's an issue getting our fingerprint
+
+    get_info_mock.side_effect = ControllerError('nope, too bad')
+
+    try:
+      self.controller.get_network_status()
+      self.fail("We should've raised an exception")
+    except ControllerError as exc:
+      self.assertEqual('Unable to determine our own fingerprint: nope, too bad', str(exc))
+
+    self.assertEqual('boom', self.controller.get_network_status(default = 'boom'))
+
+    # successful request
+
+    desc = NS_DESC % ('moria1', '/96bKo4soysolMgKn5Hex2nyFSY')
+
+    get_info_mock.side_effect = lambda param, **kwargs: {
+      'fingerprint': '9695DFC35FFEB861329B9F1AB04C46397020CE31',
+      'ns/id/9695DFC35FFEB861329B9F1AB04C46397020CE31': desc,
+    }[param]
+
+    self.assertEqual(stem.descriptor.router_status_entry.RouterStatusEntryV3(desc), self.controller.get_network_status())
+
+  @patch('stem.control.Controller.get_info')
   def test_get_network_status(self, get_info_mock):
     """
     Exercises the get_network_status() method.
@@ -389,7 +420,7 @@ class TestControl(unittest.TestCase):
 
     nickname = 'Beaver'
     fingerprint = '/96bKo4soysolMgKn5Hex2nyFSY'
-    desc = 'r %s %s u5lTXJKGsLKufRLnSyVqT7TdGYw 2012-12-30 22:02:49 77.223.43.54 9001 0\ns Fast Named Running Stable Valid\nw Bandwidth=75' % (nickname, fingerprint)
+    desc = NS_DESC % (nickname, fingerprint)
     router = stem.descriptor.router_status_entry.RouterStatusEntryV3(desc)
 
     # always return the same router status entry